% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Align_to_Bin.R
\name{Align_to_Bin}
\alias{Align_to_Bin}
\title{Align_to_Bin}
\usage{
Align_to_Bin(bed_file, bin_reference)
}
\arguments{
\item{bin_reference}{}
}
\value{

}
\description{
Align_to_Bin
}
\examples{
Align_to_Bin(bed_file, bin_reference)
}
