% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateIntervals.R
\name{annotateIntervals}
\alias{annotateIntervals}
\title{AnnotateIntervals}
\usage{
annotateIntervals(intervals, txdb)
}
\arguments{
\item{intervals}{GRange file}

\item{txdb}{gene database}
}
\value{
Annotated GRange
}
\description{
This is function is used to annotate the gene ranges
}
\examples{

annotateIntervals(GRange, hg19_txdb)

}
