% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectTranslocation.R
\name{clusterOnePairofChromosomes}
\alias{clusterOnePairofChromosomes}
\title{clusterOnePairofChromosomes}
\usage{
clusterOnePairofChromosomes(translocation_bed)
}
\arguments{
\item{translocation_bed}{dataframe, paired translocation bed dataframe
Columns:
-Chr_1 and Pos_1: First break point chromosome ID and location.
-Chr_2 and Pos_2: Second breal point chromosome ID and location.}
}
\value{
sorted translocation bed dataframe
}
\description{
Clustering on one pair of chromosomes
}
