% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genome_Windows.R
\name{Genome_Windows}
\alias{Genome_Windows}
\title{Genome Windows}
\usage{
Genome_Windows(reference = "hg19", tilewidth = 5e+05)
}
\arguments{
\item{tilewidth}{}
}
\description{
Genome Windows
}
\examples{
Genome_Windows("hg19", 500000)

}
