% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getParentFrame}
\alias{getParentFrame}
\title{Find out the parent calling function name}
\usage{
getParentFrame(parent_level = 1, char = TRUE)
}
\arguments{
\item{parent_level}{int, how many parent call stack levels to traces back,
number must bigger than 1, because 0 is \code{getParentFrame} itself.}

\item{char}{bool, convert the return to character? \code{FALSE} will be original \code{call}
object}
}
\description{
Find out the parent calling function name
}
\details{
This function must be used in another function
}
