% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hard_Filter.R
\name{Hard_Filter}
\alias{Hard_Filter}
\title{Hard Filter the VCF
Hard filter the raw vcf file based on the "PASS", "PRECISE" and "Both"}
\usage{
Hard_Filter(VCF_List, Filter_Standard)
}
\arguments{
\item{Filter_Standard}{}
}
\value{
Filtered_VCF_List
}
\description{
Hard Filter the VCF
Hard filter the raw vcf file based on the "PASS", "PRECISE" and "Both"
}
