% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hg38coord}
\alias{hg38coord}
\alias{hg38special}
\title{Extra Homo sapiens hg38 genome coordinates}
\usage{
hg38coord(muted = FALSE)

hg38special(muted = FALSE)
}
\arguments{
\item{muted}{bool, print additional explanation of columns?}
}
\value{
a tibble coordinates in bed-like format
}
\description{
\itemize{
\item \link{hg38coord} returns a tibble of all hg38 chromosome coordinates
\item \link{hg38special} returns a tibble of all hg38 special region coordinates
}
}
\examples{
hg38coord()
hg38specail()
}
