% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection_of_intervals.R
\name{intersection_bed_df}
\alias{intersection_bed_df}
\title{Intersection of intervals}
\usage{
intersection_bed_df(bed_df_1, bed_df_2, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
Intersection of intervals
}
