% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection_of_intervals.R
\name{intersectionBedDf}
\alias{intersectionBedDf}
\alias{\%I\%}
\title{Intersection of intervals}
\usage{
intersectionBedDf(bed_df_1, bed_df_2, verbose = FALSE)

bed_list_1 \%I\% bed_list_2
}
\arguments{
\item{bed_df_1}{xxx}

\item{bed_df_2}{xxx}

\item{verbose}{bool, print verbose information?}
}
\value{
xxx
}
\description{
Intersection of intervals
}
\examples{
1+1
}
