% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_of_GRange.R
\name{overlap_of_GRange}
\alias{overlap_of_GRange}
\title{Title}
\usage{
overlap_of_GRange(grl_1, grl_2)
}
\arguments{
\item{grl_1}{GRange 1}

\item{grl_2}{GRange 2}
}
\value{
The overlap of gtl_1 and grl_2
}
\description{
Title
}
\examples{
overlap_of_GRange(grl_1,grl_2)
}
