% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectTranslocation.R
\name{projectTranslocation}
\alias{projectTranslocation}
\title{Project translocation to Cartesian coordinate system}
\usage{
projectTranslocation(translocation_bed)
}
\arguments{
\item{translocation_bed}{dataframe, paired translocation bed dataframe
Columns:
-Chr_1 and Pos_1: First break point chromosome ID and location.
-Chr_2 and Pos_2: Second breal point chromosome ID and location.}
}
\value{
a ggplot object
}
\description{
Visualization of translocations in Cartesian coordinate system.
}
\examples{
translocation_bed <- data.frame(Chr_1 = paste0("chr", c(7, 8, 9, 9, 13, 15)),
Pos_1 = c(8663231, 70602300, 131457166, 33130549, 21746650, 42750778),
Chr_2 = paste0("chr", c(5, 1, 2, 6, 11, 2)),
Pos_2 = c(37709720, 91853200, 116376668, 43655549, 108585748, 214996194))

projectTranslocation(translocation_bed)

}
