% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectTranslocation.R
\name{projectTranslocationClustering}
\alias{projectTranslocationClustering}
\title{Project translocation to Cartesian coordinate system and clustering}
\usage{
projectTranslocationClustering(
  translocation_bed,
  method = "K-Means",
  least_translocation_number_of_clustering = 10
)
}
\arguments{
\item{translocation_bed}{dataframe, paired translocation bed dataframe
Columns:
-Chr_1 and Pos_1: First break point chromosome ID and location.
-Chr_2 and Pos_2: Second breal point chromosome ID and location.}

\item{method}{clustering method, now only support "K-means", will add other methods into the function}

\item{least_translocation_number_of_clustering}{numeric,
when the Chr_1 to Chr_2 translocation number > least_translocation_number_of_clustering, perform clustering,
otherwise no need to clustering}
}
\value{
clustered result
}
\description{
Project translocation to Cartesian coordinate system and clustering
}
