% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_VCFs.R
\name{Read_VCFs}
\alias{Read_VCFs}
\title{Read_VCFs}
\usage{
Read_VCFs(sample_mapping_df)
}
\arguments{
\item{sample_mapping_df}{}
}
\value{
VCF_list
}
\description{
Read_VCFs
}
\examples{

Read_VCFs(sample_mapping_df)
}
