% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_VCFs.R
\name{readVCFs}
\alias{readVCFs}
\title{readVCFs}
\usage{
readVCFs(sample_mapping_df)
}
\arguments{
\item{sample_mapping_df}{dataframe, the input sample map indicate the caller,
sample and directory.
Columns:
\itemize{
\item Sample: Sample's name.
\item Caller: Caller's name.
\item Directory: directory to the VCF file.
}}
}
\value{
VCFList
}
\description{
readVCFs
}
\examples{
\dontrun{
Read_VCFs(sample_mapping_df)
}


}
