% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setColorContinuous}
\alias{setColorContinuous}
\alias{setColorDiscrete}
\title{VCFComparison package plot colors}
\usage{
setColorContinuous(plot_colors = "current", muted = FALSE)

setColorDiscrete(plot_colors = "current", n = 8, muted = FALSE)
}
\arguments{
\item{plot_colors}{character vector, colors to use for gradients for discrete palettes.

If the \code{plot_colors} is keyword "current", only display current color options but \strong{DO NOT}
set new color options.

If the value is \code{"default"}, for \link{setColorDiscrete} it uses \emph{Set3}, for
\link{setColorContinuous} it uses \emph{plasma} gradient.}

\item{muted}{bool, muted the print message on console?}

\item{n}{integer, for \link{setColorDiscrete} only, how many unique colors to create based on
provided colors. See \link[grDevices:colorRamp]{grDevices::colorRampPalette} for details.}
}
\value{
nothing will be returned, by the plot color options will be set, use
\code{VCFComparisonOption("color_cont")} and \code{VCFComparisonOption("color_dis")}
to check current values after using setting the colors.
}
\description{
get VCFComparison package plots colors, or set custom colors for
both continuous and discrete colors.
}
\details{
\link{setColorContinuous} will set color palettes for continuous plot gradients,
\link{setColorDiscrete} set the colors to use for discrete color palettes.
}
\examples{
setColorContinuous(c("red", "blue", "yellow"))
setColorDiscrete(c("red", "blue", "yellow"), 10)
}
