% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_bed.R
\name{sort_trans_bed}
\alias{sort_trans_bed}
\title{Sort Translocation Bed data}
\usage{
sort_trans_bed(input_trans_bed)
}
\arguments{
\item{input_trans_bed}{}
}
\value{

}
\description{
Sort Translocation Bed data
}
