% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectTranslocation.R
\name{sortTranslocationBed}
\alias{sortTranslocationBed}
\title{sortTranslocationBed}
\usage{
sortTranslocationBed(input_trans_bed)
}
\arguments{
\item{translocation_bed}{dataframe, paired translocation bed dataframe
Columns:
-Chr_1 and Pos_1: First break point chromosome ID and location.
-Chr_2 and Pos_2: Second breal point chromosome ID and location.}
}
\value{
sorted translocation bed dataframe
}
\description{
Sort the translocation bed dataframe and output the translocation dataframe following Chr_1 <= Chr_2 and rows sorted from chr1 to chrY as well.
}
\examples{
translocation_bed <- data.frame(Chr_1 = paste0("chr", c(7, 8, 9, 9, 13, 15)),
Pos_1 = c(8663231, 70602300, 131457166, 33130549, 21746650, 42750778),
Chr_2 = paste0("chr", c(5, 1, 2, 6, 11, 2)),
Pos_2 = c(37709720, 91853200, 116376668, 43655549, 108585748, 214996194))

sortTranslocationBed(translocation_bed)

}
