% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_type_List.R
\name{SV_type_List}
\alias{SV_type_List}
\title{SV type extraction
This function is used to extract specific SV types from the SV list}
\usage{
SV_type_List(VCF_list, SV_type)
}
\arguments{
\item{SV_type}{}
}
\description{
SV type extraction
This function is used to extract specific SV types from the SV list
}
\examples{
SV_type_List(SVF_list, SV_type)
}
