% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sv_chr.R
\name{svChrPlot}
\alias{svChrPlot}
\title{Plot SVs by chromosome}
\usage{
svChrPlot(sv, ref = "hg38", label_karyo = TRUE)
}
\arguments{
\item{ref}{string, what genome to use, default is "hg38"}
}
\value{

}
\description{
Plot SVs by chromosome
}
