% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SVIVOption}
\alias{SVIVOption}
\title{Get or set options for SVIV package}
\usage{
SVIVOption(option = NULL, value = NULL)
}
\arguments{
\item{option}{string, what option to get or set}

\item{value}{any R object, the value you want to set for the option.}
}
\value{
see details
}
\description{
Get or set options for SVIV package
}
\details{
If value is not provided, get the current value of this option, if
non-NULL value is provided, set the option with this value. If the value is not
provided and the option value is unset, return \code{FALSE}
\subsection{Current possible values and defaults}{
\itemize{
\item verbose: FALSE
}
}
}
\examples{
SVIVOption("verbose") # get current value
SVIVOption("verbose", TRUE) # set it to a new value
SVIVOption("verbose") # check the value again
}
