% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_type_List.R
\name{SVtypeList}
\alias{SVtypeList}
\title{SV type extraction}
\usage{
SVtypeList(VCF_list, SV_type)
}
\arguments{
\item{VCF_list}{xxx}

\item{SV_type}{xxx}
}
\value{
what is the return
}
\description{
This function is used to extract specific SV types from the SV list
}
\examples{
1 + 1
# SVtypeList(VCF_list, SV_type)
}
