% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_of_intervals.R
\name{unionBedDf}
\alias{unionBedDf}
\alias{\%U\%}
\title{Union of intervals}
\usage{
unionBedDf(bed_df_1, bed_df_2, verbose = FALSE)

bed_list_1 \%U\% bed_list_2
}
\arguments{
\item{bed_df_1}{xxx}

\item{bed_df_2}{xxx}

\item{verbose}{bool, to display verbose information?}
}
\value{
xxx
}
\description{
Union of intervals
}
\examples{
1 + 1
}
