% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_circos.R
\name{vcf_circos}
\alias{vcf_circos}
\title{Circos plot from VCF file}
\usage{
vcf_circos(SV_fix)
}
\arguments{
\item{SV_fix}{}
}
\value{
circos plot
}
\description{
Circos plot from VCF file
}
\examples{
vcf_circos(VCF_file)
}
