% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hard_Filter.R
\name{hardFilter}
\alias{hardFilter}
\title{Hard Filter the VCF}
\usage{
hardFilter(VCF_list, filter_standard)
}
\arguments{
\item{filter_standard}{character, choose from "PASS", "Precise", or "Both".}

\item{VCFlist}{VCF_list, the list of input vcf files.}
}
\value{
Filtered VCF_list
}
\description{
Filter the VCF list by filter tag
}
\examples{
\dontrun{
Read_VCFs(sample_mapping_df)
}

}
