% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectTranslocation.R
\name{projectTranslocationClustering}
\alias{projectTranslocationClustering}
\title{Project translocation to Cartesian coordinate system and clustering}
\usage{
projectTranslocationClustering(
  translocation_bed,
  method = "K-Means",
  least_translocation_number_of_clustering = 10
)
}
\arguments{
\item{translocation_bed}{dataframe, paired translocation bed dataframe
Columns:
-Chr_1 and Pos_1: First break point chromosome ID and location.
-Chr_2 and Pos_2: Second breal point chromosome ID and location.}

\item{method}{clustering method, now only support "K-means", will add other methods into the function}

\item{least_translocation_number_of_clustering}{numeric,
when the Chr_1 to Chr_2 translocation number > least_translocation_number_of_clustering, perform clustering,
otherwise no need to clustering}
}
\value{
clustered result
}
\description{
Project translocation to Cartesian coordinate system and clustering
}
\examples{
\dontrun{
translocation_bed_1 <- data.frame(Chr_1 = paste0("chr", c(7, 8, 9, 9, 13, 15)),
Pos_1 = c(8663000, 70602500, 131457200, 33130600, 21746650, 42750778),
Chr_2 = paste0("chr", c(5, 1, 2, 6, 11, 2)),
Pos_2 = c(37709720, 91853200, 116376668, 43655549, 108585748, 214996194))
translocation_bed_2 <- data.frame(Chr_1 = paste0("chr", c(7, 8, 9, 9, 13, 15)),
                                  Pos_1 = c(8663000, 70602500, 131457200, 33130600, 21746650, 42750778)/100,
                                  Chr_2 = paste0("chr", c(5, 1, 2, 6, 11, 2)),
                                  Pos_2 = c(37709720, 91853200, 116376668, 43655549, 108585748, 214996194)/100)

translocation_bed_1_2 <- rbind(translocation_bed_1, translocation_bed_2)
translocation_bed <- data.frame()
for(i in 1:10){
   translocation_bed_new <- translocation_bed_1_2 \%>\% dplyr::mutate(Pos_1 = Pos_1 + sample.int(1000, 12)) \%>\% dplyr::mutate(Pos_2 = Pos_2 + sample.int(1000, 12))
   translocation_bed <- rbind(translocation_bed, translocation_bed_new)
}
projectTranslocationClustering(translocation_bed)
}
}
