% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_type_List.R
\name{SVtypeList}
\alias{SVtypeList}
\title{SVtypeList}
\usage{
SVtypeList(VCF_list, SV_type)
}
\arguments{
\item{VCF_list}{list, the list of VCF files}

\item{SV_type}{characters, one from DEL, DUP, INV, INS, and TRANS}
}
\value{
SVtypeList
}
\description{
This function is used to extract specific SV types from the SV list
}
\examples{
\dontrun{
SVtypeList(VCF_list, SV_type)
}
}
