% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translocation_gene_list.R
\name{translocation_gene_list}
\alias{translocation_gene_list}
\title{Title}
\usage{
translocation_gene_list(input_vcf_list, txdb)
}
\arguments{
\item{input_vcf_list}{vcf GRane list}

\item{txdb}{gene database}
}
\value{
translocation gene list
}
\description{
Title
}
\examples{
translocation_gene_list(vcf_list, txdb)

}
