% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_of_intervals.R
\name{unionBedDf}
\alias{unionBedDf}
\alias{\%U\%}
\title{Union of intervals}
\usage{
unionBedDf(bed_df_1, bed_df_2, verbose = FALSE)

bed_list_1 \%U\% bed_list_2
}
\arguments{
\item{bed_df_1}{dataframe, bed format dataframe noting one type of the interval mutation.
Columns:
\itemize{
\item Chr:  string, chromosome numbers example: chr1, chr2, etc
\item Start:  numeric integers, SV starting position
\item End:    numeric integers, SV ending p1osition
}}

\item{bed_df_2}{dataframe, bed format dataframe noting one type of the interval mutation.
Columns:
\itemize{
\item Chr:  string, chromosome numbers example: chr1, chr2, etc
\item Start:  numeric integers, SV starting position
\item End:    numeric integers, SV ending p1osition
}}

\item{verbose}{bool, to display verbose information?}
}
\value{
bed_df
}
\description{
Union operation on two bed dataframes.
}
\examples{
bed_df_1 <- data.frame(Chr = c(paste0("chr", c(1,1,1))), Start=c(100, 200, 300), End=c(150, 250, 350))
bed_df_2 <- data.frame(Chr = c(paste0("chr", c(1,1,1))), Start=c(120, 220, 320), End=c(150, 250, 350))
unionBedDf(bed_df_1, bed_df_2, verbose = FALSE)

}
